Struct actix_test::TestRequest

pub struct TestRequest { /* private fields */ }

Test Request builder.

For unit testing, actix provides a request builder type and a simple handler runner. TestRequest implements a builder-like pattern. You can generate various types of request via TestRequest’s methods:

TestRequest::to_request creates an actix_http::Request.

TestRequest::to_srv_request creates a [ServiceRequest], which is used for testing middlewares and chain adapters.

TestRequest::to_srv_response creates a [ServiceResponse].

TestRequest::to_http_request creates an [HttpRequest], which is used for testing handlers.

use actix_web::{test, HttpRequest, HttpResponse, HttpMessage};
use actix_web::http::{header, StatusCode};

async fn handler(req: HttpRequest) -> HttpResponse {
    if let Some(hdr) = req.headers().get(header::CONTENT_TYPE) {
        HttpResponse::Ok().into()
    } else {
        HttpResponse::BadRequest().into()
    }
}

#[actix_web::test]
async fn test_index() {
    let req = test::TestRequest::default()
        .insert_header(header::ContentType::plaintext())
        .to_http_request();

    let resp = handler(req).await;
    assert_eq!(resp.status(), StatusCode::OK);

    let req = test::TestRequest::default().to_http_request();
    let resp = handler(req).await;
    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
}

Implementations

impl TestRequest

pub fn with_uri(uri: &str) -> TestRequest

Constructs test request and sets request URI.

pub fn get() -> TestRequest

Constructs test request with GET method.

pub fn post() -> TestRequest

Constructs test request with POST method.

pub fn put() -> TestRequest

Constructs test request with PUT method.

pub fn patch() -> TestRequest

Constructs test request with PATCH method.

pub fn delete() -> TestRequest

Constructs test request with DELETE method.

pub fn version(self, ver: Version) -> TestRequest

Sets HTTP version of this request.

pub fn method(self, meth: Method) -> TestRequest

Sets method of this request.

pub fn uri(self, path: &str) -> TestRequest

Sets URI of this request.

pub fn insert_header(self, header: impl TryIntoHeaderPair) -> TestRequest

Inserts a header, replacing any that were set with an equivalent field name.

pub fn append_header(self, header: impl TryIntoHeaderPair) -> TestRequest

Appends a header, keeping any that were set with an equivalent field name.

pub fn cookie(self, cookie: Cookie<'_>) -> TestRequest

Available on crate feature cookies only.

Sets cookie for this request.

pub fn param( self, name: impl Into<Cow<'static, str>>, value: impl Into<Cow<'static, str>> ) -> TestRequest

Sets request path pattern parameter.

Examples

use actix_web::test::TestRequest;

let req = TestRequest::default().param("foo", "bar");
let req = TestRequest::default().param("foo".to_owned(), "bar".to_owned());

pub fn peer_addr(self, addr: SocketAddr) -> TestRequest

Sets peer address.

pub fn set_payload(self, data: impl Into<Bytes>) -> TestRequest

Sets request payload.

pub fn set_form(self, data: impl Serialize) -> TestRequest

Serializes data to a URL encoded form and set it as the request payload.

The Content-Type header is set to application/x-www-form-urlencoded.

pub fn set_json(self, data: impl Serialize) -> TestRequest

Serializes data to JSON and set it as the request payload.

The Content-Type header is set to application/json.

pub fn app_data<T>(self, data: T) -> TestRequest
where T: 'static,

Inserts application data.

This is equivalent of App::app_data() method for testing purpose.

pub fn to_request(self) -> Request

Finalizes request creation and returns Request instance.

pub fn to_srv_request(self) -> ServiceRequest

Finalizes request creation and returns ServiceRequest instance.

pub fn to_srv_response<B>(self, res: HttpResponse<B>) -> ServiceResponse<B>

Finalizes request creation and returns ServiceResponse instance.

pub fn to_http_request(self) -> HttpRequest

Finalizes request creation and returns HttpRequest instance.

pub fn to_http_parts(self) -> (HttpRequest, Payload)

Finalizes request creation and returns HttpRequest and Payload pair.

pub async fn send_request<S, B, E>( self, app: &S ) -> <S as Service<Request>>::Response
where S: Service<Request, Response = ServiceResponse<B>, Error = E>, E: Debug,

Finalizes request creation, calls service, and waits for response future completion.